Update LinuxI2cCommunicator.cpp
On my machine, if I ran the original code, there was garbage after the name of the file and I always got the error message. 
This could have also been fixed by making a filename_ a char[11] instead of char[10], but this also cleans up the case when the length of the bus_number is more than one. (: